Xext: move shm extension under it's own subdir
